Revert "refactor(ui): reorganize DropIndicator import path for better structure 🔄📁"
This reverts commit c3ab5ecc6d30a988ff4e8f4489b124b1432f3d35.
