+++ frontend/src/features/cards/components/ui/index.ts
@@ -2,4 +2,3 @@ export { CardPriorityBadge } from "./CardPriorityBadge";
 export { CardTypeBadge } from "./CardTypeBadge";
 export { CardDueDateDisplay } from "./CardDueDateDisplay";
 export { CardTagsDisplay } from "./CardTagsDisplay";
-export { DropIndicator } from "./DropIndicator";
